requirements.txt
python-pptx>=0.6.21
python-docx>=0.8.11
streamlit>=1.31.0
